🚚 Move non-pm task into swift package
