+++ src/main/kotlin/at/phatbl/swift/VersionTask.kt
@@ -1,9 +1,11 @@
-package at.phatbl.swiftpm.tasks
+package at.phatbl.swift
+
+import at.phatbl.swiftpm.tasks.AbstractExecTask
 
 /**
  * Prints the version of swift.
  */
-open class SwiftVersionTask : AbstractExecTask() {
+open class VersionTask : AbstractExecTask() {
     init {
         description = "Prints the version of Swift being used."
         command = "swift --version"
