+++ src/main/kotlin/at/phatbl/swiftpm/SwiftPMPlugin.kt
@@ -1,5 +1,6 @@
 package at.phatbl.swiftpm
 
+import at.phatbl.swift.VersionTask
 import at.phatbl.swiftpm.Constants.Companion.TASK_CLEAN
 import at.phatbl.swiftpm.Constants.Companion.TASK_DESCRIBE
 import at.phatbl.swiftpm.Constants.Companion.TASK_DUMP_PACKAGE
@@ -16,26 +17,28 @@ import org.gradle.api.Task
 class SwiftPMPlugin : Plugin<Project> {
     lateinit var project: Project
 
+    lateinit var swiftVersion: Task
+
     lateinit var clean: Task
     lateinit var describe: Task
     lateinit var dumpPackage: Task
     lateinit var generateXcodeTask: Task
+    lateinit var pmVersion: Task
     lateinit var reset: Task
-    lateinit var swiftVersion: Task
     lateinit var toolsVersion: Task
-    lateinit var version: Task
 
     override fun apply(project: Project) {
         this.project = project
+
+        swiftVersion = project.tasks.create(TASK_SWIFT_VERSION, VersionTask::class.java)
 
         clean = project.tasks.create(TASK_CLEAN, CleanTask::class.java)
         describe = project.tasks.create(TASK_DESCRIBE, DescribeTask::class.java)
         dumpPackage = project.tasks.create(TASK_DUMP_PACKAGE, DumpPackageTask::class.java)
         generateXcodeTask = project.tasks.create(TASK_GENERATE_XCODE_PROJECT, GenerateXcodeTask::class.java)
+        pmVersion = project.tasks.create(TASK_VERSION, PMVersionTask::class.java)
         reset = project.tasks.create(TASK_RESET, ResetTask::class.java)
-        swiftVersion = project.tasks.create(TASK_SWIFT_VERSION, SwiftVersionTask::class.java)
         toolsVersion = project.tasks.create(TASK_TOOLS_VERSION, ToolsVersionTask::class.java)
-        version = project.tasks.create(TASK_VERSION, VersionTask::class.java)
     }
 
     /**
