+++ src/main/kotlin/at/phatbl/swiftpm/tasks/PMVersionTask.kt
@@ -3,7 +3,7 @@ package at.phatbl.swiftpm.tasks
 /**
  * Prints the version of the swift package manager.
  */
-open class VersionTask : AbstractExecTask() {
+open class PMVersionTask : AbstractExecTask() {
     init {
         description = "Prints the version of SwiftPM being used."
         command = "swift package --version"
