+++ src/test/kotlin/at/phatbl/swift/VersionTaskSpek.kt
@@ -1,6 +1,6 @@
-package at.phatbl.swiftpm.tasks
+package at.phatbl.swift
 
-import at.phatbl.swiftpm.Constants.Companion.TASK_VERSION
+import at.phatbl.swiftpm.Constants.Companion.TASK_SWIFT_VERSION
 import org.gradle.testfixtures.ProjectBuilder
 import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
@@ -9,9 +9,9 @@ import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
 object VersionTaskSpek: Spek({
-    describe("version task") {
+    describe("swift version task") {
         val project = ProjectBuilder.builder().build()
-        val task = project.tasks.create(TASK_VERSION, VersionTask::class.java)
+        val task = project.tasks.create(TASK_SWIFT_VERSION, VersionTask::class.java)
 
         it("can be added to project") {
             assertNotNull(task)
@@ -21,7 +21,6 @@ object VersionTaskSpek: Spek({
         it("has a tokenized command line") {
             val expectedTokens = arrayOf(
                 "swift",
-                "package",
                 "--version"
             )
